Move some schema to Question.ts
